Add files via upload
This code is divided into 4 parts :

Identify the Shorter Array: Start by identifying the array with the minimum length. This helps in reducing the number of operations since we can use the smaller array to build the hash table.

Count Occurrences Using Hash Tables: Use hash tables (or dictionaries in Python) to count the occurrences of each element in both arrays.

Compare Occurrences: Compare the occurrences of each element found in the shorter array with the occurrences in the longer array.

Find Common Elements: The common elements are those that have non-zero counts in both hash tables.